adding condition to check xargo available on host before running docker, adding multiple terminal for debug (not yet tested it, only gnoe-temrinal tested) and splitting srcs to files.mk

+++ Makefile
@@ -4,6 +4,18 @@ QEMU			=	qemu-system-i386
 
 HOST			=	$(shell uname)
 
+ifneq ($(shell which gnome-terminal),)
+TERM_EMU	=	gnome-terminal --
+else ifneq ($(shell which xterm),)
+TERM_EMU	=	xterm -e
+else ifneq ($(shell which konsole),)
+TERM_EMU	=	konsole -e
+else ifeq ($(shell which terminator),)
+TERM_EMU	=	terminator -x
+else
+TERM_EMU	=	$(TERM_EMU)
+endif
+
 TARGER_ARCH 	=	i386
 
 LINKERFILE		=	linker.ld
@@ -31,24 +43,7 @@ DIR_ISO			=	iso
 DIR_GRUB		=	$(DIR_ISO)/boot/grub
 
 vpath %.s $(foreach dir, ${shell find $(DIR_SRCS) -type d}, $(dir))
-
-BOOTSRCS		=	header.s \
-					boot.s \
-					gdt.s
-
-BOOTOBJS		=	$(BOOTSRCS:%.s=$(DIR_OBJS)/%.o)
-
-RUST_SRCS		=	main.rs \
-					mod.rs \
-					io.rs \
-					keyboard.rs \
-					cursor.rs \
-					color.rs \
-					gdt.rs \
-					cli.rs
-
-KERNELSRCS		=	$(foreach file, $(RUST_SRCS), $(shell find $(DIR_SRCS) -name $(file) -type f))
-
+include files.mk
 
 RUST_KERNEL 	=	target/i386-kfs/debug/libkernel.a
 NAME			=	kfs_$(VERSION)
@@ -58,38 +53,48 @@ all:			$(NAME)
 boot:			$(NAME)
 				$(QEMU) -d int -drive format=raw,file=$(NAME) -serial file:$(MAKEFILE_PATH)kernel.log 2> qemu.log
 
+# This rule will run qemu with flags to wait gdb to connect to it
 debug:			$(NAME)
 				$(QEMU) -s -S -daemonize -drive format=raw,file=$(NAME) -serial file:$(MAKEFILE_PATH)kernel.log
-				terminator -x "cd Documents/kfs; gdb $(DIR_ISO)/boot/$(NAME)"
+				$(TERM_EMU) bash -c "cd $(MAKEFILE_PATH); gdb $(DIR_ISO)/boot/$(NAME) -x gdbstart"
 
+test:
+	echo $(TEST)
+
+# Rule to create iso file which can be run with qemu
 $(NAME):		$(DIR_ISO)/boot/$(NAME) $(DIR_GRUB)/$(GRUB_CFG)
 ifeq ($(shell docker images -q ${DOCKER_GRUB} 2> /dev/null),)
 				docker build $(DOCKER_DIR) -f $(DOCKER_DIR)/$(DOCKER_GRUB).dockerfile -t $(DOCKER_GRUB)
 endif
-				docker run --rm -v $(MAKEFILE_PATH):/root:Z $(DOCKER_GRUB) -o $(NAME) $(DIR_ISO)
+				docker run --rm -u $(shell id -u ${USER}):$(shell id -g ${USER}) -v $(MAKEFILE_PATH):/root:Z $(DOCKER_GRUB) -o $(NAME) $(DIR_ISO)
 
+# Link asm file with rust according to the linker script in arch directory
 $(DIR_ISO)/boot/$(NAME):		$(BOOTOBJS) $(RUST_KERNEL) | $(DIR_GRUB)
 ifeq ($(shell docker images -q ${DOCKER_LINKER} 2> /dev/null),)
 				docker build $(DOCKER_DIR) -f $(DOCKER_DIR)/$(DOCKER_LINKER).dockerfile -t $(DOCKER_LINKER)
 endif
-				docker run --rm -v $(MAKEFILE_PATH):/root:Z $(DOCKER_LINKER) $(LINKERFLAGS) $^ -o $(DIR_ISO)/boot/$(NAME)
+				docker run --rm -u $(shell id -u ${USER}):$(shell id -g ${USER}) -v $(MAKEFILE_PATH):/root:Z $(DOCKER_LINKER) $(LINKERFLAGS) $^ -o $(DIR_ISO)/boot/$(NAME)
 
 $(DIR_GRUB):
 				mkdir -p $(DIR_GRUB)
 
+# Build libkernel using xargo
 $(RUST_KERNEL):	$(KERNELSRCS)
 ifeq ($(shell docker images -q ${DOCKER_RUST} 2> /dev/null),)
 				docker build $(DOCKER_DIR) -f $(DOCKER_DIR)/$(DOCKER_RUST).dockerfile -t $(DOCKER_RUST)
 endif
+ifeq ($(shell which xargo),)
 				docker run --rm -v $(MAKEFILE_PATH):/root:Z $(DOCKER_RUST) build --target=$(TARGER_ARCH)-kfs
+else
+				xargo build --target $(TARGER_ARCH)-kfs
+endif
 
+# Check if the rust can compile without actually compiling it
 check: $(KERNELSRCS)
 ifeq ($(shell docker images -q ${DOCKER_RUST} 2> /dev/null),)
 				docker build $(DOCKER_DIR) -f $(DOCKER_DIR)/$(DOCKER_RUST).dockerfile -t $(DOCKER_RUST)
 endif
 				docker run -t --rm -v $(MAKEFILE_PATH):/root:Z $(DOCKER_RUST) check
-
-
 
 $(DIR_GRUB)/$(GRUB_CFG): $(DIR_CONFIG)/$(GRUB_CFG)
 				cp -f $(DIR_CONFIG)/$(GRUB_CFG) $(DIR_GRUB)
